Refactored voor makkelijker kopieeren.
